refactor(waterfall): 获取瀑布图的默认配置项
