--- src/plots/waterfall/constants.ts
@@ -1,4 +1,0 @@
-export const Y_FIELD = '$$yField$$';
-export const DIFF_FIELD = '$$diffField$$';
-export const ABSOLUTE_FIELD = '$$absoluteField$$';
-export const IS_TOTAL = '$$isTotal$$';
